+++ packages/@expo/metro-config/src/serializer/fork/__tests__/mini-metro.test.ts
@@ -71,6 +71,7 @@ it(`can create a micro Metro graph fixture`, async () => {
                   },
                   "hasCjsExports": false,
                   "lineCount": 6,
+                  "loaderReference": undefined,
                   "map": [
                     [
                       4,
@@ -226,6 +227,7 @@ it(`can create a micro Metro graph fixture`, async () => {
                   },
                   "hasCjsExports": false,
                   "lineCount": 14,
+                  "loaderReference": undefined,
                   "map": [
                     [
                       7,