+++ packages/expo-video/src/VideoView.web.tsx
@@ -63,9 +63,6 @@ export const VideoView = forwardRef((props: { player?: VideoPlayer } & VideoView
 
   useImperativeHandle(ref, () => ({
     enterFullscreen: async () => {
-      if (!props.allowsFullscreen || !videoRef.current) {
-        return;
-      }
       // Cast the video to any to avoid ts errors. Methods such as webkitRequestFullscreen,
       // webkitEnterFullScreen, msRequestFullscreen are not typed even though they exist.
       const video = videoRef.current as any;
@@ -263,7 +260,7 @@ export const VideoView = forwardRef((props: { player?: VideoPlayer } & VideoView
   return (
     <video
       controls={props.nativeControls ?? true}
-      controlsList={props.allowsFullscreen ? undefined : 'nofullscreen'}
+      controlsList={props.fullscreenOptions?.enable ? undefined : 'nofullscreen'}
       crossOrigin={props.crossOrigin}
       style={{
         ...mapStyles(props.style),